refactor: remove nested utils module
