+++ gateway-common/src/http_ext.rs
