+++ gateway-common/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod blocklist;
 pub mod caching;
+pub mod http_ext;
 pub mod ptr;
+pub mod timestamp;
 pub mod ttl_hash_map;
-pub mod utils;
