--- gateway-common/src/utils/mod.rs
@@ -1,4 +1,0 @@
-pub mod http_ext;
-pub mod testing;
-pub mod timestamp;
-pub mod tracing;
